refactor(eslint): centralize TypeScript rule utilities
